Move get_fasta_hdas_by_history into services
